BE/BACKEND/requirements.txt
asgiref
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
psycopg2
python-dotenv